refactoring. interface BankStatementParser BankStatementProcessor for class BankStatementCSVParser pattern strategy.
